+++ changelog.md
@@ -1,4 +1,5 @@
 ### 4.0.0-alpha5
+* Add `/lwc details <HistoryId>` which allows you to view specific information on a history item.
 * Add `/lwc history` - used to view history items for a specific player or all latest ones. To be used hand in hand with `/lwc details` which views specific information on any given history item. The arguments for `/lwc history` are dynamic and as such, all of the following uses are valid: `/lwc history` `/lwc history *` `/lwc history 2` (page 2) `/lwc history 2 *` `/lwc history * 2` `/lwc history Hidendra` `/lwc history Hidendra`  ... and so on!
 * The way chests were facing were broken when upgrading from a pre-1.8 map to 1.8. The first time you click a protected chest you can access, it will reorient itself. Alternatively, you can use `/lwc fix` after this to manually orient a chest.
 * **FIX:** Limits will now use the highest group limit instead of the first one found.